requirements.txt
aniso8601==9.0.1
anyio==3.7.1
click==8.1.6
exceptiongroup==1.1.2
fastapi==0.95.1
graphene==3.2.2
graphql-core==3.2.3
graphql-relay==3.2.0
h11==0.14.0
idna==3.4
psycopg==3.1.8
pydantic==1.10.12
sniffio==1.3.0
SQLAlchemy==2.0.11
starlette==0.26.1
starlette-graphene3==0.6.0
typing_extensions==4.7.1
uvicorn==0.21.1
